requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
chromadb>=0.5.0
gradio>=4.0
pydantic>=2.0
pyyaml>=6.0
python-dotenv>=1.0
sentence-transformers>=2.2.0
Pillow>=10.0
